Break up tool validator into Galaxy/TS variants.
